requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
pydantic[email]
pydantic-settings
httpx
bcrypt
PyJWT
python-multipart